requirements.txt
ipykernel
numpy
matplotlib
scipy
Pillow
opencv-python
joblib
scikit-learn
pandas
seaborn
torch
torchvision
torchsummary
pyserial
hidapi
#torch torchvision --index-url https://download.pytorch.org/whl/cu118
